Divided project into 2 different projects targetting .NET classic and portable platforms.
