+++ unirest-net/packages/repositories.config
@@ -1,5 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <repositories>
-  <repository path="..\unirest-net\packages.config" />
-  <repository path="..\unirest-net-tests\packages.config" />
+  <repository path="..\unirest-net (net40)\packages.config" />
+  <repository path="..\unirest-net (portable)\packages.config" />
+  <repository path="..\unirest-net-tests (net40)\packages.config" />
 </repositories>
