+++ unirest-net/unirest-net (net40)/packages.config
@@ -1,0 +1,7 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Microsoft.Bcl" version="1.1.6" targetFramework="net40" />
+  <package id="Microsoft.Bcl.Build" version="1.0.13" targetFramework="net40" />
+  <package id="Microsoft.Net.Http" version="2.2.15" targetFramework="net40" />
+  <package id="Newtonsoft.Json" version="5.0.8" targetFramework="portable-net40+sl50+wp80+win" />
+</packages>
