+++ unirest-net/unirest-net.sln
@@ -1,10 +1,13 @@
 ﻿
-Microsoft Visual Studio Solution File, Format Version 11.00
-# Visual Studio 2010
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 2013
+VisualStudioVersion = 12.0.21005.1
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "unirest-net", "unirest-net\unirest-net.csproj", "{BB1927F2-4FC7-4520-8680-49C0245E05E0}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "unirest-net (net40)", "unirest-net (net40)\unirest-net (net40).csproj", "{BB1927F2-4FC7-4520-8680-49C0245E05E0}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "unirest-net-tests", "unirest-net-tests\unirest-net-tests.csproj", "{A1455397-D38D-4855-AEF6-634EE9DCD10E}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "unirest-net (portable)", "unirest-net (portable)\unirest-net (portable).csproj", "{B3CAC40B-7582-4058-8F65-197301EE6807}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "unirest-net-tests (net40)", "unirest-net-tests (net40)\unirest-net-tests (net40).csproj", "{A1455397-D38D-4855-AEF6-634EE9DCD10E}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -16,6 +19,10 @@ Global
 		{BB1927F2-4FC7-4520-8680-49C0245E05E0}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{BB1927F2-4FC7-4520-8680-49C0245E05E0}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{BB1927F2-4FC7-4520-8680-49C0245E05E0}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B3CAC40B-7582-4058-8F65-197301EE6807}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B3CAC40B-7582-4058-8F65-197301EE6807}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B3CAC40B-7582-4058-8F65-197301EE6807}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B3CAC40B-7582-4058-8F65-197301EE6807}.Release|Any CPU.Build.0 = Release|Any CPU
 		{A1455397-D38D-4855-AEF6-634EE9DCD10E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{A1455397-D38D-4855-AEF6-634EE9DCD10E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A1455397-D38D-4855-AEF6-634EE9DCD10E}.Release|Any CPU.ActiveCfg = Release|Any CPU
